Remove unused tabs
